requirements.txt
Flask==2.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
openpyxl>=3.1.0

